rag_khotba/requirements.txt
streamlit
torch
sentence-transformers
faiss-cpu
google-cloud-texttospeech
pandas
numpy
scikit-learn
transformers
streamlit-aggrid